Listener organization better way to organize listeners, use their name to tell what they do
