+++ src/main/java/net/pixlies/events/listeners/ListenerManager.java
@@ -2,8 +2,8 @@ package net.pixlies.events.listeners;
 
 import lombok.Getter;
 import net.pixlies.Main;
-import net.pixlies.events.listeners.chat.ChatListener;
-import net.pixlies.events.listeners.join.LoginListener;
+import net.pixlies.events.listeners.moderation.MuteListener;
+import net.pixlies.events.listeners.moderation.BanListener;
 import org.bukkit.Bukkit;
 import org.bukkit.event.Listener;
 
@@ -15,8 +15,8 @@ public class ListenerManager {
     private static final Main instance = Main.getInstance();
 
     private @Getter static final List<Listener> listeners = new ArrayList<>(){{
-        add(new LoginListener());
-        add(new ChatListener());
+        add(new BanListener());
+        add(new MuteListener());
     }};
 
     private ListenerManager() {}
